backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
langchain>=0.1.0
langchain-openai>=0.0.8
openai>=1.10.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
numpy>=1.26.0
scikit-learn>=1.3.0
